Move data writing logic to JsonFactory class
