--- data/data.json
@@ -1,14 +1,0 @@
-{
-  "audio" : [ {
-    "title" : "Thriller",
-    "id" : "5920",
-    "status" : "AVAILABLE",
-    "loanee" : null,
-    "dueDate" : null,
-    "artist" : "Michael Jackson",
-    "duration" : "42:16"
-  } ],
-  "book" : [ ],
-  "magazine" : [ ],
-  "video" : [ ]
-}
